# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Standout Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /brand:
    get:
      operationId: getBrandProfile
      tags: [brand]
      summary: Get user's brand profile
      responses:
        "200":
          description: Brand profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandProfile"
    put:
      operationId: updateBrandProfile
      tags: [brand]
      summary: Update brand profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandProfileBody"
      responses:
        "200":
          description: Updated brand profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandProfile"
  /brand/reflections:
    get:
      operationId: getBrandReflections
      tags: [brand]
      summary: Get brand reflection prompts
      responses:
        "200":
          description: Reflection prompts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReflectionPrompt"
    post:
      operationId: submitReflection
      tags: [brand]
      summary: Submit a reflection response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReflectionBody"
      responses:
        "200":
          description: Reflection saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List user AI coach conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new AI coach conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpenaiConversationBody"
      responses:
        "201":
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"
  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get a conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
  /openai/conversations/{id}/messages:
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send message to AI coach (streaming SSE)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenaiMessageBody"
      responses:
        "200":
          description: Streaming SSE response
          content:
            text/event-stream:
              schema:
                type: string
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    BrandProfile:
      type: object
      properties:
        desiredIdentity:
          type: array
          items:
            type: string
        currentPerceptions:
          type: array
          items:
            type: string
        brandScore:
          type: number
        gapAreas:
          type: array
          items:
            type: string
        strengths:
          type: array
          items:
            type: string
        habits:
          type: array
          items:
            $ref: "#/components/schemas/BrandHabit"
        lastUpdated:
          type: string
          nullable: true
      required: [desiredIdentity, currentPerceptions, brandScore, gapAreas, strengths, habits]
    BrandHabit:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        isActive:
          type: boolean
        streakDays:
          type: integer
      required: [id, title, description, isActive, streakDays]
    BrandProfileBody:
      type: object
      properties:
        desiredIdentity:
          type: array
          items:
            type: string
        currentPerceptions:
          type: array
          items:
            type: string
        habitIds:
          type: array
          items:
            type: string
    ReflectionPrompt:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        category:
          type: string
        isAnswered:
          type: boolean
        answer:
          type: string
          nullable: true
      required: [id, text, category, isAnswered]
    ReflectionBody:
      type: object
      properties:
        promptId:
          type: string
        answer:
          type: string
      required: [promptId, answer]
...
    OpenaiConversation:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        messageCount:
          type: integer
      required: [id, title, createdAt, updatedAt, messageCount]
    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required: [id, title, createdAt, updatedAt, messages]
    OpenaiMessage:
      type: object
      properties:
        id:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
      required: [id, role, content, createdAt]
    CreateOpenaiConversationBody:
      type: object
      properties:
        title:
          type: string
      required: [title]
    SendOpenaiMessageBody:
      type: object
      properties:
        content:
          type: string
        toneMode:
          type: string
          enum: [diplomatic, firm, executive_ready, empathetic, politically_careful]
      required: [content]